cleanup
minor

move files into libs

map values
