+++ libs/expo/betterangels/src/lib/screens/Client/ClientProfile_V2/ClientProfileSections/ImportantNotes.tsx
@@ -1,7 +1,7 @@
 import { TextRegular } from '@monorepo/expo/shared/ui-components';
 import { View } from 'react-native';
+import { ClientProfileInfo } from '../../../../ui-components';
 import { ClientProfileQuery } from '../../__generated__/Client.generated';
-import { ClientProfileSection } from '../ClientProfileSection/ClientProfileSection';
 
 type TProps = {
   client: ClientProfileQuery | undefined;
@@ -20,12 +20,11 @@ export default function ImportantNotes(props: TProps) {
 
   return (
     <View>
-      <ClientProfileSection
+      <ClientProfileInfo
         items={content}
-        showAll
         action={{
           onClick: () => alert('clicked'),
-          accessibilityLabel: 'edit name information',
+          accessibilityLabel: 'edit important notes',
         }}
       />
     </View>
