+++ libs/expo/betterangels/src/lib/ui-components/ClientProfileInfo/index.ts
@@ -1,0 +1,1 @@
+export { ClientProfileInfo } from './ClientProfileInfo';
