+++ libs/expo/betterangels/src/lib/ui-components/InfoList/index.ts
@@ -1,0 +1,2 @@
+export { InfoList } from './InfoList';
+export { InfoListItem, TInfoListItem } from './InfoListItem';
