+++ libs/expo/betterangels/src/lib/ui-components/index.ts
@@ -4,6 +4,7 @@ export { default as Attachments } from './Attachments';
 export { default as AuthContainer } from './AuthContainer';
 export { default as ClientCard } from './ClientCard';
 export { default as ClientCardModal } from './ClientCardModal';
+export { ClientProfileInfo } from './ClientProfileInfo';
 export { default as ConsentModal } from './ConsentModal';
 export { default as DateOfBirthPicker } from './DateOfBirthPicker';
 export { default as DocumentModal } from './DocumentModal';
@@ -12,6 +13,7 @@ export { FileThumbnail } from './FileThumbnail/FileThumbnail';
 export { default as GenderPicker } from './GenderPicker';
 export { default as HairColorPicker } from './HairColorPicker';
 export { default as Header } from './Header';
+export { InfoList, InfoListItem, TInfoListItem } from './InfoList';
 export { default as LoginForm } from './LoginForm';
 export { default as MainContainer } from './MainContainer';
 export { default as MainPlusModal } from './MainPlusModal';
